+++ src/main/java/nedelosk/modularmachines/common/blocks/item/ItemBlockModularAssembler.java
@@ -1,11 +1,6 @@
 package nedelosk.modularmachines.common.blocks.item;
 
-import nedelosk.forestday.common.machines.mutiblock.charcoalkiln.TileCharcoalKiln;
-import nedelosk.forestday.common.machines.mutiblock.charcoalkiln.WoodType;
-import nedelosk.forestday.common.registrys.FBlocks;
-import nedelosk.modularmachines.common.blocks.ModularBlock;
 import nedelosk.modularmachines.common.blocks.tile.TileModularAssembler;
-import nedelosk.modularmachines.common.core.BlockRegistry;
 import nedelosk.modularmachines.common.core.MMBlocks;
 import net.minecraft.block.Block;
 import net.minecraft.entity.player.EntityPlayer;
@@ -86,7 +81,7 @@ public class ItemBlockModularAssembler extends ItemBlock {
 			TileModularAssembler assembler = (TileModularAssembler) tile;
 			assembler.setCapacity(stack.getTagCompound().getInteger("capacity"));
 			world.markBlockForUpdate(x, y, z);
-			world.playSoundEffect((double)((float)x + 0.5F), (double)((float)y + 0.5F), (double)((float)z + 0.5F), this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
+			world.playSoundEffect(x + 0.5F, y + 0.5F, z + 0.5F, this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
 			stack.stackSize--;
 
 			return true;
